+++ Application/Project_IdehService/Query/GetProject_IdehById.cs
@@ -84,11 +84,11 @@ namespace Application.Project_IdehService.Query
         public bool JustificationPlan { get; set; }
         public int IndustryId { get; set; }
         public string IndustryName { get; set; }
-        public int RequiredCapitalPlan { get; set; }
-        public int AmountCapitalDemand { get; set; }
+        public long RequiredCapitalPlan { get; set; }
+        public long AmountCapitalDemand { get; set; }
         public string WhatTopicsNeedParticipate { get; set; }
         public int ReturnInvestmentRate { get; set; }
-        public int NetPresentValue { get; set; }
+        public long NetPresentValue { get; set; }
         public DateTime TimeCreate { get; set; }
         public AddressDetailDto Address { get; set; }
         public List<Link> Links { get; set; }
