+++ Application/Project_ReadyToUseService/Command/AddProject_ReadyToUseService.cs
@@ -100,7 +100,7 @@ namespace Application.Project_ReadyToUseService.Command
         //نرخ بازگشت سرمایه طرح
         public int ReturnInvestmentRate { get; set; }
         //میزان سرمایه مورد تقاضا
-        public int AmountCapitalDemand { get; set; }
+        public long AmountCapitalDemand { get; set; }
         //آدرس
         public CreateAddressDto Address { get; set; }
         /// <summary>
@@ -126,15 +126,15 @@ namespace Application.Project_ReadyToUseService.Command
         public bool JustificationPlan { get; set; }
         //واگذاری
         //قیمت پیشنهادی
-        public int ProposedPrice { get; set; }
+        public long ProposedPrice { get; set; }
         //مشارکت
         //کل سرمایه مورد نیاز طرح
-        public int RequiredCapitalPlan { get; set; }
+        public long RequiredCapitalPlan { get; set; }
         //مشارکت
         //در چه سرفصل‌هایی نیاز به مشارکت دارید؟
         public string WhatTopicsNeedParticipate { get; set; }
         //خالص ارزش فعلی پروژه
-        public int NetPresentValue { get; set; }
+        public long NetPresentValue { get; set; }
         //درصد پیشرفت کلی طرح تا کنون
         public int OverallProgressPercent { get; set; }
         //آیا تاکنون در خصوص طرح مذکور تسهیلات دریافت نموده‌اید؟
