+++ WebApi/ModelsAndDtoes/Project_Development/CreateProject_DevelopmentApiDto.cs
@@ -19,7 +19,7 @@ namespace WebApi.ModelsAndDtoes.Project_Development
         //نرخ بازگشت سرمایه طرح
         public int ReturnInvestmentRate { get; set; }
         //میزان سرمایه مورد تقاضا
-        public int AmountCapitalDemand { get; set; }
+        public long AmountCapitalDemand { get; set; }
         //آدرس
         [Required]
         public CreateAddressApiDto Address { get; set; }
