+++ WebApi/ModelsAndDtoes/Project_Stopped/CreateProject_StoppedApiDto.cs
@@ -17,7 +17,7 @@ namespace WebApi.ModelsAndDtoes.Project_Stopped
         //نرخ بازگشت سرمایه طرح
         public int ReturnInvestmentRate { get; set; }
         //میزان سرمایه مورد تقاضا
-        public int AmountCapitalDemand { get; set; }
+        public long AmountCapitalDemand { get; set; }
         //آدرس
         public CreateAddressDto Address { get; set; }
         /// <summary>
@@ -57,7 +57,7 @@ namespace WebApi.ModelsAndDtoes.Project_Stopped
         //ارزش کل سرمایه گذاری جدید
         public long TotalValueNewInvestment { get; set; }
         //خالص ارزش فعلی پروژه
-        public int NetPresentValue { get; set; }
+        public long NetPresentValue { get; set; }
         //دلایل غیر فعال بودن طرح
         public int InactivityReasonId { get; set; }
         //توضیح
